Fix warning due to type conversion in SendGMTicketGetTicket

+++ src/game/GMTickets/GMTicketHandler.cpp
@@ -27,7 +27,7 @@
 
 void WorldSession::SendGMTicketGetTicket(uint32 status, GMTicket* ticket /*= nullptr*/) const
 {
-    int len = ticket ? strlen(ticket->GetText()) : 0;
+    size_t len = ticket ? strlen(ticket->GetText()) : 0;
     WorldPacket data(SMSG_GMTICKET_GETTICKET, (4 + len + 1 + 4 + 2 + 4 + 4));
     data << uint32(status);                                 // standard 0x0A, 0x06 if text present
     if (status == 6)
